adapt object model
