+++ src/components/dashboard/immunizationStatus/immunizationCardConfigurations.ts
@@ -1,4 +1,4 @@
-import { AggregatedImmunizationStatus } from '../../../core/models/aggregatedImmunizationStatus';
+import { AggregatedImmunizationStatus } from '../../../core/models/AggregatedImmunizationStatus';
 import {
   FaExclamationCircle,
   FaInfoCircle,
