+++ src/core/models/ActiveVaccinationScheme.ts
@@ -1,0 +1,5 @@
+
+export interface ActiveVaccinationScheme {
+  changeReason: string;
+  vaccinationSchemeId: string;
+}
