--- src/core/models/ActiveVaccinationSchme.ts
@@ -1,6 +1,0 @@
-import { VaccinationScheme } from './VaccinationScheme';
-
-export interface ActiveVaccinationScheme {
-  changeReason: string;
-  vaccinationScheme: VaccinationScheme;
-}
