+++ src/core/models/Address.ts
@@ -1,0 +1,8 @@
+export interface Address {
+    city: string;
+    country: string;
+    district: string;
+    line: string;
+    postalCode: string;
+    state: string;
+}
