+++ src/core/models/AggregatedImmunizationStatus.ts
