+++ src/core/models/Disease.ts
@@ -4,4 +4,6 @@ export interface Disease {
   code: CodeableConcept;
   name: string;
   description: string;
+  populationRecommendationId: string;
+  vaccineIds: string[];
 }
