+++ src/core/models/Gender.ts
@@ -1,0 +1,1 @@
+export type Gender = 'male' | 'female' | 'other' | 'unknown';
