+++ src/core/models/HumanName.ts
@@ -1,0 +1,6 @@
+export interface HumanName {
+    family: string;
+    given: string[];
+    prefix: string[];
+    suffix: string[];
+}
