--- src/core/models/MedicationIngredient.ts
@@ -1,6 +1,0 @@
-import { CodeableConcept } from './CodeableConcept';
-
-export interface MedicationIngredient {
-    itemCodeableConcept: CodeableConcept;
-    itemReference: string; // // id refers to a Substance id
-}
