+++ src/core/models/PopulationRecommendation.ts
@@ -2,4 +2,5 @@ export interface PopulationRecommendation {
   ageStart: number;
   ageEnd: number;
   location: Location;
+  diseaseId: string;
 }
