+++ src/core/models/VaccinationDose.ts
@@ -1,0 +1,16 @@
+export interface VaccinationDose {
+  doseQuantity: number;
+  isProtected: boolean;
+  notes: string;
+  vaccinationSchemeId: string;
+}
+
+export interface VaccinationDoseSingle extends VaccinationDose {
+  numberInScheme: number;
+  timeStart: Date;
+  timeEnd: Date;
+}
+
+export interface VaccinationDoseRepeating extends VaccinationDose {
+  interval: Array<number>;
+}
