--- src/core/models/VaccinationDosis.ts
@@ -1,15 +1,0 @@
-export interface VaccinationDosis {
-  doseQuantity: number;
-  isProtected: boolean;
-  notes: string;
-}
-
-export interface VaccinationDosisSingle extends VaccinationDosis {
-  numberInScheme: number;
-  timeStart: Date;
-  timeEnd: Date;
-}
-
-export interface VaccinationDosisRepeating {
-  interval: Array<number>;
-}
